style: don't skip string normalization using Black

+++ ibis/backends/impala/tests/test_pandas_interop.py
@@ -17,8 +17,8 @@ pytest.importorskip("impala")
 def exhaustive_df():
     return pd.DataFrame(
         {
-            'bigint_col': np.array([0, 10, 20, 30, 40, 50, 60, 70, 80, 90], dtype='i8'),
-            'bool_col': np.array(
+            "bigint_col": np.array([0, 10, 20, 30, 40, 50, 60, 70, 80, 90], dtype="i8"),
+            "bool_col": np.array(
                 [
                     True,
                     False,
@@ -33,19 +33,19 @@ def exhaustive_df():
                 ],
                 dtype=np.bool_,
             ),
-            'date_string_col': [
-                '11/01/10',
+            "date_string_col": [
+                "11/01/10",
                 None,
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
-                '11/01/10',
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
+                "11/01/10",
             ],
-            'double_col': np.array(
+            "double_col": np.array(
                 [
                     0.0,
                     10.1,
@@ -60,7 +60,7 @@ def exhaustive_df():
                 ],
                 dtype=np.float64,
             ),
-            'floatcol': np.array(
+            "floatcol": np.array(
                 [
                     np.nan,
                     1.1000000238418579,
@@ -73,37 +73,37 @@ def exhaustive_df():
                     8.8000001907348633,
                     9.8999996185302734,
                 ],
-                dtype='f8',
+                dtype="f8",
             ),
-            'int_col': np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype='i4'),
-            'month': [11, 11, 11, 11, 2, 11, 11, 11, 11, 11],
-            'smallint_col': np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype='i2'),
-            'string_col': [
-                '0',
-                '1',
+            "int_col": np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype="i4"),
+            "month": [11, 11, 11, 11, 2, 11, 11, 11, 11, 11],
+            "smallint_col": np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype="i2"),
+            "string_col": [
+                "0",
+                "1",
                 None,
-                'double , whammy',
-                '4',
-                '5',
-                '6',
-                '7',
-                '8',
-                '9',
+                "double , whammy",
+                "4",
+                "5",
+                "6",
+                "7",
+                "8",
+                "9",
             ],
-            'timestamp_col': [
-                pd.Timestamp('2010-11-01 00:00:00'),
+            "timestamp_col": [
+                pd.Timestamp("2010-11-01 00:00:00"),
                 None,
-                pd.Timestamp('2010-11-01 00:02:00.100000'),
-                pd.Timestamp('2010-11-01 00:03:00.300000'),
-                pd.Timestamp('2010-11-01 00:04:00.600000'),
-                pd.Timestamp('2010-11-01 00:05:00.100000'),
-                pd.Timestamp('2010-11-01 00:06:00.150000'),
-                pd.Timestamp('2010-11-01 00:07:00.210000'),
-                pd.Timestamp('2010-11-01 00:08:00.280000'),
-                pd.Timestamp('2010-11-01 00:09:00.360000'),
+                pd.Timestamp("2010-11-01 00:02:00.100000"),
+                pd.Timestamp("2010-11-01 00:03:00.300000"),
+                pd.Timestamp("2010-11-01 00:04:00.600000"),
+                pd.Timestamp("2010-11-01 00:05:00.100000"),
+                pd.Timestamp("2010-11-01 00:06:00.150000"),
+                pd.Timestamp("2010-11-01 00:07:00.210000"),
+                pd.Timestamp("2010-11-01 00:08:00.280000"),
+                pd.Timestamp("2010-11-01 00:09:00.360000"),
             ],
-            'tinyint_col': np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype='i1'),
-            'year': [
+            "tinyint_col": np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype="i1"),
+            "year": [
                 2010,
                 2010,
                 2010,
@@ -168,7 +168,7 @@ def test_insert(con, temp_table_db, exhaustive_df):
 
     table = con.table(table_name, database=tmp_db)
 
-    result = table.execute().sort_values(by='tinyint_col').reset_index(drop=True)
+    result = table.execute().sort_values(by="tinyint_col").reset_index(drop=True)
     tm.assert_frame_equal(result, exhaustive_df)
 
 
@@ -186,8 +186,8 @@ def _check_roundtrip(con, df):
 
 
 def test_timestamp_with_timezone():
-    df = pd.DataFrame({'A': pd.date_range('20130101', periods=3, tz='US/Eastern')})
+    df = pd.DataFrame({"A": pd.date_range("20130101", periods=3, tz="US/Eastern")})
     schema = sch.infer(df)
-    expected = ibis.schema([('A', "timestamp('US/Eastern')")])
+    expected = ibis.schema([("A", "timestamp('US/Eastern')")])
     assert schema.equals(expected)
-    assert schema.types[0].equals(dt.Timestamp('US/Eastern'))
+    assert schema.types[0].equals(dt.Timestamp("US/Eastern"))
